requirements.txt
newspaper3k==0.2.8
jsonpickle==1.1.0
simplejson==3.16.0